refactor: folder restructure CF-552
