+++ synapsepy/__init__.py
@@ -1,6 +1,5 @@
-from .client import Client
-from .user import User, Users
-from .node import Node, Nodes
-from .subnet import Subnet, Subnets
-from .subscription import Subscription, Subscriptions
-from .transaction import Trans, Transactions
+from lib.client import Client
+from lib.user import User, Users
+
+
+print("Success")
