+++ synapsepy/helpers/errors.py
