+++ synapsepy/lib/client.py
@@ -1,13 +1,13 @@
 
-from .http_client import HttpClient
+from helpers.http_client import HttpClient
 
-from .user import User, Users
-from .node import Node, Nodes
-from .transaction import Trans, Transactions
-from .subscription import Subscription, Subscriptions
-from . import errors as api_errors
+from lib.user import User, Users
+from helpers.node import Node, Nodes
+from helpers.transaction import Trans, Transactions
+from helpers.subscription import Subscription, Subscriptions
+import helpers.errors as api_errors
 
-from .endpoints import paths
+from constants.endpoints import paths
 
 import sys
 import json
