+++ synapsepy/lib/user.py
@@ -1,9 +1,9 @@
 
-from .endpoints import paths
+from constants.endpoints import paths
 
-from .node import Node, Nodes
-from .subnet import Subnet, Subnets
-from .transaction import Trans, Transactions
+from helpers.node import Node, Nodes
+from helpers.subnet import Subnet, Subnets
+from helpers.transaction import Trans, Transactions
 
 from functools import partial
 
@@ -11,7 +11,7 @@ import json
 import logging
 import requests
 import warnings
-from . import errors as api_errors
+import helpers.errors as api_errors
 
 class User():
 	""" User Record
